requirements.txt
# Core dependencies
python-dotenv==1.0.1
pydantic==2.6.1

# Google Gemini
google-generativeai==0.4.1

# LangGraph and LangChain
langgraph==0.0.26
langchain-core==0.1.27
langchain-google-genai==0.0.11

# MetaTrader 5 (Windows only)
MetaTrader5>=5.0.45

# Technical Analysis
pandas==2.2.0
numpy==1.26.4
ta==0.11.0

# Database
aiosqlite==0.19.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
tenacity==8.2.3
structlog==24.1.0

# GUI
PyQt6==6.6.1
pyqtgraph==0.13.3

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-mock>=3.12.0

# Optional: Redis for distributed state
# redis==5.0.1
